suppress compiler warnings (about halfway through)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2096 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/libs/luafontforge/fontforge/fontforge/fvcomposit.c
@@ -128,7 +128,7 @@ return( uni );
 
     for ( j=0; accents[j][0]!=0xffff; ++j ) {
 	for ( k=0; k<4 && accents[j][k]!=0; ++k ) {
-	    if ( uni==accents[j][k] ) {
+	    if ( uni==(int)accents[j][k] ) {
 		uni = 0x300+j;
 	break;
 	    }
